// src/plugins/ingredient/ingredient.types.ts
export type InputMaybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
  Upload: any;
};
// ...
export type CreateIngredientInput = {
  addedSugars: Scalars['Float'];
  calcium: Scalars['Float'];
  calories: Scalars['Float'];
  cholesterol: Scalars['Float'];
  fiber: Scalars['Float'];
  iron: Scalars['Float'];
  name: Scalars['String'];
  potassium: Scalars['Float'];
  protein: Scalars['Float'];
  saturatedFat: Scalars['Float'];
  sodium: Scalars['Float'];
  totalCarbs: Scalars['Float'];
  totalFat: Scalars['Float'];
  totalSugars: Scalars['Float'];
  transFat: Scalars['Float'];
  vitaminD: Scalars['Float'];
};
// ...
export type DateOperators = {
  after?: InputMaybe<Scalars['DateTime']>;
  before?: InputMaybe<Scalars['DateTime']>;
  between?: InputMaybe<DateRange>;
  eq?: InputMaybe<Scalars['DateTime']>;
  isNull?: InputMaybe<Scalars['Boolean']>;
};
// ...
export type IdOperators = {
  eq?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<Scalars['String']>>;
  isNull?: InputMaybe<Scalars['Boolean']>;
  notEq?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<Scalars['String']>>;
};
// ...
export type IngredientFilterParameter = {
  addedSugars?: InputMaybe<NumberOperators>;
  calcium?: InputMaybe<NumberOperators>;
  calories?: InputMaybe<NumberOperators>;
  cholesterol?: InputMaybe<NumberOperators>;
  createdAt?: InputMaybe<DateOperators>;
  fiber?: InputMaybe<NumberOperators>;
  id?: InputMaybe<IdOperators>;
  iron?: InputMaybe<NumberOperators>;
  name?: InputMaybe<StringOperators>;
  potassium?: InputMaybe<NumberOperators>;
  protein?: InputMaybe<NumberOperators>;
  saturatedFat?: InputMaybe<NumberOperators>;
  sodium?: InputMaybe<NumberOperators>;
  totalCarbs?: InputMaybe<NumberOperators>;
  totalFat?: InputMaybe<NumberOperators>;
  totalSugars?: InputMaybe<NumberOperators>;
  transFat?: InputMaybe<NumberOperators>;
  updatedAt?: InputMaybe<DateOperators>;
  vitaminD?: InputMaybe<NumberOperators>;
};

export type IngredientList = PaginatedList & {
  __typename?: 'IngredientList';
  items: Array<Ingredient>;
  totalItems: Scalars['Int'];
};

export type IngredientListOptions = {
  /** Allows the results to be filtered */
  filter?: InputMaybe<IngredientFilterParameter>;
  /** Specifies whether multiple "filter" arguments should be combines with a logical AND or OR operation. Defaults to AND. */
  filterOperator?: InputMaybe<LogicalOperator>;
  skip?: InputMaybe<Scalars['Int']>;
  /** Specifies which properties to sort the results by */
  sort?: InputMaybe<IngredientSortParameter>;
  take?: InputMaybe<Scalars['Int']>;
};

export type IngredientSortParameter = {
  addedSugars?: InputMaybe<SortOrder>;
  calcium?: InputMaybe<SortOrder>;
  calories?: InputMaybe<SortOrder>;
  cholesterol?: InputMaybe<SortOrder>;
  createdAt?: InputMaybe<SortOrder>;
  fiber?: InputMaybe<SortOrder>;
  id?: InputMaybe<SortOrder>;
  iron?: InputMaybe<SortOrder>;
  name?: InputMaybe<SortOrder>;
  potassium?: InputMaybe<SortOrder>;
  protein?: InputMaybe<SortOrder>;
  saturatedFat?: InputMaybe<SortOrder>;
  sodium?: InputMaybe<SortOrder>;
  totalCarbs?: InputMaybe<SortOrder>;
  totalFat?: InputMaybe<SortOrder>;
  totalSugars?: InputMaybe<SortOrder>;
  transFat?: InputMaybe<SortOrder>;
  updatedAt?: InputMaybe<SortOrder>;
  vitaminD?: InputMaybe<SortOrder>;
};
// ...
export enum LogicalOperator {
  AND = 'AND',
  OR = 'OR'
}
// ...
export type NumberOperators = {
  between?: InputMaybe<NumberRange>;
  eq?: InputMaybe<Scalars['Float']>;
  gt?: InputMaybe<Scalars['Float']>;
  gte?: InputMaybe<Scalars['Float']>;
  isNull?: InputMaybe<Scalars['Boolean']>;
  lt?: InputMaybe<Scalars['Float']>;
  lte?: InputMaybe<Scalars['Float']>;
};
// ...
export enum SortOrder {
  ASC = 'ASC',
  DESC = 'DESC'
}
// ...
export type StringOperators = {
  contains?: InputMaybe<Scalars['String']>;
  eq?: InputMaybe<Scalars['String']>;
  in?: InputMaybe<Array<Scalars['String']>>;
  isNull?: InputMaybe<Scalars['Boolean']>;
  notContains?: InputMaybe<Scalars['String']>;
  notEq?: InputMaybe<Scalars['String']>;
  notIn?: InputMaybe<Array<Scalars['String']>>;
  regex?: InputMaybe<Scalars['String']>;
};